requirements.txt
asyncio
pillow
pyyaml
pyocr